logical/aw: move sts signing request to awsutil
